Add time zone choice to about page
