+++ templates/index.html
@@ -193,7 +193,7 @@
                 <h3>{{ html("timezone-title") }}</h3>
                 <p>
                   {{ html("timezone-description") }}
-                  <span id="timezone-info">{{ html("loading") }}</span><br/>
+                  <span class="timezone-of-browser">{{ html("loading") }}</span><br/>
                   <select id="select-timezone">
                       <option value="">{{ html("timezone-browser") }}</option>
                   </select>
