+++ translations/en/about.yml
@@ -8,11 +8,20 @@ description: "Here is more information about the calendar. Have a look at the li
 # https://open-web-calendar.hosted.quelltext.eu/about.html#translate-website
 links-website: "Go back to view the calendar."
 # One of the links
-# https://open-web-calendar.hosted.quelltext.eu/about.html#translate-website
+# https://open-web-calendar.hosted.quelltext.eu/about.html#translate-links-ics
 links-ics: "Use this link to subscribe to the calendar or to add the events to your own calendar."
 # One of the links
-# https://open-web-calendar.hosted.quelltext.eu/about.html#translate-website
+# https://open-web-calendar.hosted.quelltext.eu/about.html#translate-links-spec
 links-spec: "View or download the calendar specification."
+# One of the items in the menu to choose the time zone of the calendar
+# https://open-web-calendar.hosted.quelltext.eu/about.html#translate-timezone-calendar
+timezone-calendar: ""
+# One of the items in the menu to choose the time zone of the calendar
+# https://open-web-calendar.hosted.quelltext.eu/about.html#translate-timezone-browser
+timezone-browser: "This is the time zone of your device:"
+# One of the items in the menu to choose the time zone of the calendar
+# https://open-web-calendar.hosted.quelltext.eu/about.html#translate-timezone-choice
+timezone-choice: "You can choose to change the time zone of the calendar here:"
 
 
 # This is the headline/title on the about page
@@ -23,5 +32,4 @@ project-title: "About the Open Web Calendar Project"
 # Title for the license section
 # https://open-web-calendar.hosted.quelltext.eu/about.html#translate-license
 license: "License"
-
